+++ config/urls.py
@@ -28,5 +28,6 @@ urlpatterns = [
     path('', root_redirect),
     path('admin/', admin.site.urls),
     path('inventory_items/', include('items_management.urls'), name='api'),
+    path('reports/', include('reports.urls')),
     path('account/', include('account.urls'), name='account')
 ]
